feat(clients): add luajit to the list of available clients
### Summary

Adds `resty.ada` client to the list of available clients.
`resty.ada` contains LuaJIT FFI bindings to Ada URL that
is usable with LuaJIT and OpenResty projects.

At this point I didn't create Lua C API, but if I decide
so in the future, I would like to donate the client to Ada
URL project. Then it would work with PuC Lua too. I just
don't have enough time to focus on it at the moment.

Signed-off-by: Aapo Talvensaari <aapo.talvensaari@gmail.com>